fix broken app paths list
The code for detecting app paths on gnu/linux scans this list:
```emacs-lisp
(list consult-omni-apps-xdg-data-home
       consult-omni-apps-xdg-data-dirs
       "/usr/share"
       "/usr/local/share")
```
However, `consult-omni-apps-xdg-data-dirs` is a list itself and not a
string, so this will create a list with three strings and one list of
strings, which breaks the code and doesn't let any application in
`consult-omni-apps-xdg-data-dirs` be launched. To fix this, the
three strings must be appended to the `consult-omni-apps-xdg-data-dirs` string.